dd-57 Move tabs to tab package
