+++ app/src/main/java/io/krugosvet/dailydish/android/dagger/AppComponent.kt
@@ -5,7 +5,7 @@ import io.krugosvet.dailydish.android.dagger.module.AccountModule
 import io.krugosvet.dailydish.android.dagger.module.AppModule
 import io.krugosvet.dailydish.android.dagger.module.NetworkModule
 import io.krugosvet.dailydish.android.mainScreen.MealListAdapter
-import io.krugosvet.dailydish.android.mainScreen.MealListPageFragment
+import io.krugosvet.dailydish.android.mainScreen.tab.MealListPageFragment
 import io.krugosvet.dailydish.android.mainScreen.StartupActivity
 import io.krugosvet.dailydish.android.utils.baseUi.BaseActivity
 import javax.inject.Singleton
