--- app/src/main/java/io/krugosvet/dailydish/android/mainScreen/ForTodayFragment.kt
@@ -1,12 +1,0 @@
-package io.krugosvet.dailydish.android.mainScreen
-
-import android.os.Bundle
-
-class ForTodayFragment : MealListPageFragment() {
-
-    companion object {
-        fun newInstance(pageTitle: String) = ForTodayFragment().apply {
-            arguments = Bundle().apply { putString(PAGE_TITLE, pageTitle) }
-        }
-    }
-}
