+++ app/src/main/res/values/strings.xml
@@ -26,5 +26,7 @@
     <string name="sign_out">Sign out</string>
     <string name="sign_up">Sign up</string>
     <string name="sign_in">Sign in</string>
-  <string name="layout_empty_text">No meals here?\nTry to add some.</string>
+    <string name="layout_empty_text">No meals here?\nTry to add some.</string>
+    <!--suppress CheckTagEmptyBody -->
+    <string name="empty_string"></string>
 </resources>
