+++ app/src/main/res/values/styles.xml
@@ -29,7 +29,7 @@
         <item name="android:textColor">@color/gray_text</item>
     </style>
 
-    <style name="MainScreenTabs" parent="TextAppearance.Design.Tab">
+    <style name="MainScreenTabs" parent="TextAppearance.Design.Tab  ">
         <item name="textAllCaps">false</item>
         <item name="android:textAllCaps">false</item>
     </style>
